+++ src/index.ts
@@ -8,6 +8,7 @@ export { default as Select } from '@/package/components/select.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as Switch } from '@/package/components/switch.vue';
 export { default as Popover } from '@/package/components/popover.vue';
+export { default as Modal } from '@/package/components/modal.vue';
 // config
 export { default as VueForge, setTheme, updateTheme, getTheme } from '@/package/config/index';
 export type { ThemeTokens, ThemeOptions, ThemePreset } from '@/package/config/theme-core';